+++ index.d.ts
@@ -1,9 +1,17 @@
-export default initI18N;
 /**
- * @param {string} dirLocale
- * @param {Array<string>} [segmentsDirLocale]
+ *
+ * @param {string} namespace
+ * @param {string} dirResource
+ * @param {string[]} [locales]
  */
-declare function initI18N(dirLocale: string, segmentsDirLocale?: Array<string>): Promise<{
-    T: (key: import("i18next").TFunctionKeys, options: import("i18next").TOptions, lng: string) => import("i18next").TFunctionDetailedResult<object>;
-    TT: (locale: string) => (key: import("i18next").TFunctionKeys, options: import("i18next").TOptions) => import("i18next").TFunctionDetailedResult<object>;
-}>;
+export function loadI18NResource(namespace: string, dirResource: string, locales?: string[]): Promise<void>;
+export function T(key: import("i18next").TFuncKey, options: import("i18next").TOptions, lng: string): any;
+/**
+ * @param {string} namespace
+ * @param {string[]} [locales]
+ * @param {string[]} [formats]
+ * @returns {TranslatorLocale}
+ */
+export function TT(namespace: string, locales?: string[], formats?: string[]): TranslatorLocale;
+export type Translator = (key: import("i18next").TFuncKey, options: import("i18next").TOptions, lng: string) => any;
+export type TranslatorLocale = (key: import("i18next").TFuncKey, options: import("i18next").TOptions) => any;