+++ index.js
@@ -1,61 +1,87 @@
 import { readFileSync } from 'fs';
 import { resolve } from 'path';
 
-import I18N from 'i18next';
-
-import { copyJSON } from '@nuogz/utility';
-
 
 
 /**
- * @param {string} dirLocale
- * @param {Array<string>} [segmentsDirLocale]
+ * @callback Translator
+ * @param {import("i18next").TFuncKey} key
+ * @param {import("i18next").TOptions} options
+ * @param {string} lng
  */
-const initI18N = async (dirLocale, segmentsDirLocale) => {
-	const useFormatHades = (process.env.OUTPUT_FORMAT ?? '').split(';').filter(s => s).map(s => s.trim().toLowerCase()).includes('hades');
-	const language = process.env.OUTPUT_LOCALE ?? 'en';
+
+/**
+ * @callback TranslatorLocale
+ * @param {import("i18next").TFuncKey} key
+ * @param {import("i18next").TOptions} options
+ */
 
 
-	if(!segmentsDirLocale) { segmentsDirLocale = useFormatHades ? ['locale', 'hades'] : ['locale']; }
+
+// eslint-disable-next-line no-undef
+const globalTop = (globalThis ?? global ?? window);
 
 
-	const i18n = I18N.createInstance();
+const localesDefault = (globalTop?.process?.env?.NENV_I18N_LOCALE + ';en')
+	.replace(/(^|;)\s*en\s*(?=;\s*en\s*$)/, '').split(';').filter(s => s.trim()).map(s => s.trim().toLowerCase());
+const formatsDefault = (globalTop?.process?.env?.NENV_I18N_FORMAT + ';proto')
+	.replace(/(^|;)\s*proto\s*(?=;\s*proto\s*$)/, '').split(';').filter(s => s.trim()).map(s => s.trim().toLowerCase());
 
-	await i18n.init({
-		lng: language,
-		fallbackLng: 'en',
-		resources: {
-			en: JSON.parse(readFileSync(resolve(dirLocale, ...segmentsDirLocale, 'en.json'))),
-			zh: JSON.parse(readFileSync(resolve(dirLocale, ...segmentsDirLocale, 'zh.json')))
-		},
+
+if(!('NI18N' in globalTop)) {
+	/** @type {import("i18next").i18n} */
+	const NI18N = globalTop.NI18N = (await import('i18next')).default;
+
+	NI18N.init({
+		lng: localesDefault[0],
+		fallbackLng: localesDefault,
+		resources: {},
 	});
 
-	if(useFormatHades) {
-		i18n.services.formatter.add('hadesValue', value => `~{${value}}`);
-		i18n.services.formatter.add('hadesTerm', value => `~[${value}]`);
-	}
+
+	NI18N.services.formatter.add('hadesValue', value => `~{${value}}`);
+	NI18N.services.formatter.add('hadesTerm', value => `~[${value}]`);
+}
 
 
-	/**
-	 * @param {import("i18next").TFunctionKeys} key
-	 * @param {import("i18next").TOptions} options
-	 * @param {string} lng
-	 */
-	const T = (key, options = {}, lng) => i18n.t(key, Object.assign(copyJSON(options), { lng }));
-	/**
-	 * @param {string} locale
-	 */
-	const TT = locale => {
-		/**
-		 * @param {import("i18next").TFunctionKeys} key
-		 * @param {import("i18next").TOptions} options
-		 */
-		return (key, options) => T(key, options, locale);
-	};
+/**
+ *
+ * @param {string} namespace
+ * @param {string} dirResource
+ * @param {string[]} [locales]
+ */
+export async function loadI18NResource(namespace, dirResource, locales = localesDefault) {
+	/** @type {import("i18next").i18n} */
+	const NI18N = globalTop.NI18N;
 
 
-	return { T, TT };
-};
+	locales.map(locale => {
+		try {
+			const resource = JSON.parse(readFileSync(resolve(dirResource, `${locale}.json`), 'utf-8'));
+			NI18N.addResourceBundle(locale, namespace, resource, true, true);
+		}
+		catch(error) {
+			void 0;
+		}
+	});
+}
 
 
-export default initI18N;
+/** @type {Translator} */
+export function T(key, options = {}, lngs) {
+	return globalTop.NI18N.t(key, Object.assign(JSON.parse(JSON.stringify(options)), { lngs }));
+}
+
+/**
+ * @param {string} namespace
+ * @param {string[]} [locales]
+ * @param {string[]} [formats]
+ * @returns {TranslatorLocale}
+ */
+export function TT(namespace, locales, formats = formatsDefault) {
+	return (key, options) => T(
+		formats.map(format => `${namespace}:${key}_${format}`),
+		options,
+		locales,
+	);
+}